requirements.txt
fastapi
uvicorn
pymongo
openai
python-dotenv
SpeechRecognition 

